+++ src/components/GoogleMapContainer.ts
@@ -1,12 +1,13 @@
 import { Component, createElement } from "react";
-import { Location, Map, heightUnitType, widthUnitType } from "./Map";
+import { Map, heightUnitType, widthUnitType } from "./Map";
 import { Alert } from "./Alert";
-import { UrlHelper } from "../UrlHelper";
+import { ValidateConfigs } from "../utils/ValidateConfigs";
+import { Location, StaticLocation, getStaticMarkerUrl, parseStaticLocations, parseStyle } from "../utils/ContainerUtils";
 
 interface WrapperProps {
     "class"?: string;
     friendlyId: string;
-    mxObject?: mendix.lib.MxObject;
+    mxObject: mendix.lib.MxObject;
     style: string;
 }
 
@@ -37,13 +38,6 @@ interface GoogleMapContainerProps extends WrapperProps {
     zoomLevel: number;
 }
 
-interface StaticLocation {
-    address: string;
-    latitude: string;
-    longitude: string;
-    icon: string;
-}
-
 type DataSource = "static" | "context" | "XPath" | "microflow";
 
 class GoogleMapContainer extends Component<GoogleMapContainerProps, { alertMessage?: string, locations: Location[] }> {
@@ -52,7 +46,7 @@ class GoogleMapContainer extends Component<GoogleMapContainerProps, { alertMessa
     constructor(props: GoogleMapContainerProps) {
         super(props);
 
-        const alertMessage = GoogleMapContainer.validateProps(props);
+        const alertMessage = ValidateConfigs.validate(props);
         this.subscriptionHandles = [];
         this.state = { alertMessage, locations: [] };
         this.subscribe(this.props.mxObject);
@@ -79,7 +73,7 @@ class GoogleMapContainer extends Component<GoogleMapContainerProps, { alertMessa
                 optionScroll: this.props.optionScroll,
                 optionStreetView: this.props.optionStreetView,
                 optionZoomControl: this.props.optionZoomControl,
-                style: GoogleMapContainer.parseStyle(this.props.style),
+                style: parseStyle(this.props.style),
                 width: this.props.width,
                 widthUnit: this.props.widthUnit,
                 zoomLevel: this.props.zoomLevel
@@ -93,63 +87,13 @@ class GoogleMapContainer extends Component<GoogleMapContainerProps, { alertMessa
     }
 
     componentDidMount() {
-        if (!this.state.alertMessage) this.fetchData(this.props.mxObject);
+        if (!this.state.alertMessage) {
+            this.fetchData(this.props.mxObject);
+        }
     }
 
     componentWillUnmount() {
         this.subscriptionHandles.forEach(window.mx.data.unsubscribe);
-    }
-
-    public static validateProps(props: GoogleMapContainerProps) {
-        let message = "";
-        if (props.dataSource === "static" && !props.staticLocations.length) {
-            message = "At least one static location is required for 'Data source 'Static'";
-        }
-        if (props.dataSource === "static") {
-            const invalidLocations = props.staticLocations.filter(location =>
-                !location.address && !(location.latitude && location.longitude)
-            );
-            if (invalidLocations.length > 0) {
-                message = "The 'Address' or 'Latitude' and 'Longitude' "
-                    + "is required for this 'Static' data source";
-            }
-        }
-        if (props.dataSource === "XPath" && !props.locationsEntity) {
-            message = "The 'Locations entity' is required for 'Data source' 'XPath'";
-        }
-        if (props.dataSource === "microflow" && !props.dataSourceMicroflow) {
-            message = "A 'Microflow' is required for 'Data source' 'Microflow'";
-        }
-        if (props.dataSource !== "static" && (!props.addressAttribute &&
-            !(props.longitudeAttribute && props.latitudeAttribute))) {
-            message = "The 'Address attribute' or 'Latitude Attribute' and 'Longitude attribute' "
-                + "is required for this data source";
-        }
-        if (!props.autoZoom && props.zoomLevel < 2) {
-            message = "Zoom level must be greater than 1";
-        }
-
-        return message;
-    }
-
-    // Mendix does not support negative and decimal number as static inputs, so they are strings.
-    public static parseStaticLocations(props: GoogleMapContainerProps): Location[] {
-        return props.staticLocations.map(location => ({
-            address: location.address,
-            latitude: location.latitude.trim() !== "" ? Number(location.latitude) : undefined,
-            longitude: location.longitude.trim() !== "" ? Number(location.longitude) : undefined,
-            url:  GoogleMapContainer.getStaticMarkerUrl(location, props.defaultMakerIcon)
-        }));
-    }
-
-    private static getStaticMarkerUrl(location: StaticLocation, defaultMakerIcon: string): string {
-        if (location.icon) {
-            return UrlHelper.getStaticResourceUrl(location.icon);
-        }else if (defaultMakerIcon) {
-            return UrlHelper.getStaticResourceUrl(defaultMakerIcon);
-        }else {
-            return "";
-        }
     }
 
     private subscribe(contextObject?: mendix.lib.MxObject) {
@@ -175,7 +119,7 @@ class GoogleMapContainer extends Component<GoogleMapContainerProps, { alertMessa
 
     private fetchData(contextObject?: mendix.lib.MxObject) {
         if (this.props.dataSource === "static") {
-            this.setState({ locations: GoogleMapContainer.parseStaticLocations(this.props) });
+            this.setState({ locations: parseStaticLocations(this.props) });
         } else if (this.props.dataSource === "context") {
             this.fetchLocationsByContext(contextObject);
         } else if (this.props.dataSource === "XPath" && this.props.locationsEntity) {
@@ -250,33 +194,9 @@ class GoogleMapContainer extends Component<GoogleMapContainerProps, { alertMessa
         const imageKey: string = mxObject.get(this.props.markerImageAttribute) as string;
         const image = this.props.markerImages.find(value => value.enumKey === imageKey);
 
-        if (imageKey) {
-            return image
-                ? UrlHelper.getStaticResourceUrl(image.enumImage)
-                : "";
-        }else if (this.props.defaultMakerIcon) {
-            return UrlHelper.getStaticResourceUrl(this.props.defaultMakerIcon);
-        }else {
-            return "";
-        }
-    }
-
-    private static parseStyle = (style = ""): {[key: string]: string} => { // Doesn't support a few stuff.
-        try {
-            return style.split(";").reduce<{[key: string]: string}>((styleObject, line) => {
-                const pair = line.split(":");
-                if (pair.length === 2) {
-                    const name = pair[0].trim().replace(/(-.)/g, match => match[1].toUpperCase());
-                    styleObject[name] = pair[1].trim();
-                }
-                return styleObject;
-            }, {});
-        } catch (error) {
-            // tslint:disable-next-line no-console
-            window.console.log("Failed to parse style", style, error);
-        }
-
-        return {};
+        return image
+            ? getStaticMarkerUrl(image.enumImage as string, this.props.defaultMakerIcon)
+            : "";
     }
 }
