+++ src/utils/UrlHelper.ts
@@ -1,21 +1,23 @@
 export class UrlHelper {
 
     static getStaticResourceUrlFromPath(path: string) {
-        if (this.startsWith(path, "data:")) return path;
+        if (this.startsWith(path, "data:")) {
+            return path;
+        }
         // Static resource path should already contain cachebust from the modeler.
         return mx.appUrl + path;
     }
 
     static getStaticResourceUrl(url: string) {
-        if (this.startsWith(url, "data:")) return url;
+        const cacheBust = mx.server.getCacheBust();
 
+        if (this.startsWith(url, "data:")) {
+            return url;
+        }
         if (!/^\w+:\/\//.test(url)) {
             url = this.getStaticResourceUrlFromPath(url);
         }
-
-        const cacheBust = mx.server.getCacheBust();
-
-        // Only add a cache bust if it's not already there and it's a Mendix url
+        // Only add a cache bust if it's not already there and it's a Mendix URL
         if (this.startsWith(url, mx.appUrl) && !this.endsWith(url, cacheBust)) {
             url += (/\?/.test(url) ? "&" : "?") + cacheBust;
         }
@@ -29,13 +31,15 @@ export class UrlHelper {
             "changedDate=" + changedDate
         ].join("&");
 
-        if (isThumbnail) url += "&thumb=true";
+        if (isThumbnail) {
+            url += "&thumb=true";
+        }
 
         return url;
     }
 
     static getDynamicResourceUrl(guid: string, changedDate: number, isThumbnail = false) {
-        return mx.remoteUrl + this.getDynamicResourcePath(guid, changedDate, isThumbnail);
+        return `${mx.remoteUrl} ${this.getDynamicResourcePath(guid, changedDate, isThumbnail)}`;
     }
 
     private static startsWith(searchString: string, prefix: string) {
